Rename tests/core/matrix_slice.rs to tests/core/matrix_view.rs
