+++ tests/core/mod.rs
@@ -4,7 +4,7 @@ mod conversion;
 mod edition;
 mod empty;
 mod matrix;
-mod matrix_slice;
+mod matrix_view;
 #[cfg(feature = "mint")]
 mod mint;
 #[cfg(feature = "rkyv-serialize-no-std")]
